+++ backend/src/main/java/ru/compito/taskmanager/controller/TemplateController.java
@@ -3,9 +3,15 @@ package ru.compito.taskmanager.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import ru.compito.taskmanager.config.ServiceConstants;
+import ru.compito.taskmanager.entity.Task;
 import ru.compito.taskmanager.entity.TaskTemplate;
+import ru.compito.taskmanager.service.ContentRelatedRoleService;
+import ru.compito.taskmanager.service.TaskService;
 import ru.compito.taskmanager.service.TaskTemplateService;
 
 import java.util.List;
@@ -16,33 +22,62 @@ public class TemplateController {
 
     @Autowired
     private TaskTemplateService taskTemplateService;
+    @Autowired
+    private ContentRelatedRoleService contentRelatedRoleService;
+    @Autowired
+    private TaskService taskService;
 
     @GetMapping(value = "/", produces = MediaType.APPLICATION_JSON_VALUE)
-    public List<TaskTemplate> getAll() {
-        return taskTemplateService.findAll();
+    public @ResponseBody ResponseEntity<Object> getAll() {
+        List<TaskTemplate> taskTemplates = taskTemplateService.findAll();
+        return new ResponseEntity<>(taskTemplates, HttpStatus.OK);
     }
 
     @GetMapping(value = "/{taskTemplateId}/", produces = MediaType.APPLICATION_JSON_VALUE)
-    public TaskTemplate getOne(@PathVariable Integer taskTemplateId) {
-        return taskTemplateService.getOne(taskTemplateId);
+    public @ResponseBody ResponseEntity<TaskTemplate> getOne(@PathVariable Integer taskTemplateId) {
+        TaskTemplate taskTemplate = taskTemplateService.getOne(taskTemplateId);
+        return new ResponseEntity<>(taskTemplate, HttpStatus.OK);
     }
 
     @PostMapping(value = "/",
             consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public TaskTemplate create(@RequestBody TaskTemplate taskTemplate) {
-        return taskTemplateService.save(taskTemplate);
+    public @ResponseBody ResponseEntity<TaskTemplate> create(@RequestBody TaskTemplate taskTemplate) {
+        TaskTemplate newTaskTemplate = taskTemplateService.save(taskTemplate);
+        return new ResponseEntity<>(newTaskTemplate, HttpStatus.CREATED);
+
+
     }
 
     @PutMapping(value = "/{taskTemplateId}/",
             consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public TaskTemplate update(@PathVariable Integer taskTemplateId, @RequestBody TaskTemplate updatedTaskTemplate) {
-        return taskTemplateService.update(taskTemplateId,updatedTaskTemplate);
+    public @ResponseBody ResponseEntity<TaskTemplate> update(@PathVariable Integer taskTemplateId, @RequestBody TaskTemplate updatedTaskTemplate) {
+        Task task  = taskService.getByTaskTemplateId(taskTemplateId);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Integer boardId = task.getBoard().getId();
+        if(contentRelatedRoleService.isContentOwner(boardId,authentication) ||
+                contentRelatedRoleService.isContentAdministrator(boardId,authentication)||
+                contentRelatedRoleService.isContentModerator(boardId,authentication)) {
+            TaskTemplate taskTemplate = taskTemplateService.update(taskTemplateId,updatedTaskTemplate);
+            return new ResponseEntity<>(taskTemplate, HttpStatus.OK);
+        }else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @DeleteMapping("/{taskTemplateId}/")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void delete(@PathVariable Integer taskTemplateId) {
-        taskTemplateService.delete(taskTemplateId);
+    public @ResponseBody ResponseEntity<?> delete(@PathVariable Integer taskTemplateId) {
+        Task task  = taskService.getByTaskTemplateId(taskTemplateId);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Integer boardId = task.getBoard().getId();
+        if(contentRelatedRoleService.isContentOwner(boardId,authentication) ||
+                contentRelatedRoleService.isContentAdministrator(boardId,authentication)||
+                contentRelatedRoleService.isContentModerator(boardId,authentication)) {
+            taskTemplateService.delete(taskTemplateId);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
